feat: support function type for limit option
Allow `limit` option to accept a function similar to Vite's `assetsInlineLimit`,
enabling custom asset processing logic based on file path and content.

- Add function type support: `(filePath: string, content: Buffer) => boolean | undefined`
- Add `shouldProcess` utility to handle both number and function limit types
- Update type definitions and documentation
- Add playground example

Closes #119
